pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "resume-ai-multi-agent"
version = "0.1.0"
description = "Simple multi-agent RAG system for customized resumes"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26.0",
  "PyYAML>=6.0.0",
  "ollama>=0.4.0",
  "sentence-transformers>=3.0.0",
  "pypdf>=4.0.0",
  "python-docx>=1.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
resume-ai = "resume_ai.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
